requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-eventgrid
azure-storage-blob
python-dotenv
semantic-kernel
azure-ai-documentintelligence
pyodbc